+++ src/main/java/com/codeborne/selenide/impl/SelenideElementProxy.java
@@ -32,6 +32,7 @@ import static java.util.Arrays.asList;
 @ParametersAreNonnullByDefault
 class SelenideElementProxy implements InvocationHandler {
   private static final Set<String> methodsToSkipLogging = new HashSet<>(asList(
+      "as",
       "toWebElement",
       "toString",
       "getSearchCriteria",
@@ -73,7 +74,7 @@ class SelenideElementProxy implements InvocationHandler {
 
     long timeoutMs = getTimeoutMs(method, arguments);
     long pollingIntervalMs = getPollingIntervalMs(method, arguments);
-    SelenideLog log = SelenideLogger.beginStep(webElementSource.getSearchCriteria(), method.getName(), args);
+    SelenideLog log = SelenideLogger.beginStep(webElementSource.description(), method.getName(), args);
     try {
       Object result = dispatchAndRetry(timeoutMs, pollingIntervalMs, proxy, method, args);
       SelenideLogger.commitStep(log, PASS);